IDK a bunch of stuff... still figuring out what to do with this language. Scope is a bit too big right now.

+++ src/types.rs
@@ -1,1 +1,1 @@
-pub type Integer = i32;
+pub type Int = i64;
